+++ Semana_4/node/src/modules/appointments/infra/Typeorm/repositories/AppointmentsRepository.ts
@@ -1,7 +1,9 @@
 import Appointments from "../entities/Appointments";
-import { getRepository, Repository } from "typeorm";
+import { getRepository, Repository, Raw } from "typeorm";
 import IAppointmentsRepository from "@appointments/repositories/IAppointmentsRepository";
 import ICreateAppointmentsDTO from "@appointments/Dtos/ICreateAppointmentsDTO";
+import IFindAllInMonthFromProviderDTO from "@appointments/Dtos/IFindAllInMonthFromProviderDTO";
+import IFindAllInDayFromProviderDTO from "@modules/appointments/Dtos/IFindAllInDayFromProviderDTO";
 
 // O repository é responsavel pela comunicacao da aplicacao com o seu componente de pesrsistencia
 
@@ -16,6 +18,46 @@ class AppointmentsRepository implements IAppointmentsRepository {
     this.ormRepository = getRepository(Appointments); // criando Repo
   }
 
+  public async findAllInDayFromProvider({
+    day,
+    month,
+    provider_id,
+    year,
+  }: IFindAllInDayFromProviderDTO): Promise<Appointments[]> {
+    let parsedDay = String(day).padStart(2, "0");
+    let parsedMonth = String(month).padStart(2, "0");
+
+    let appointments = await this.ormRepository.find({
+      where: {
+        provider_id,
+        date: Raw(
+          (dateFieldName) =>
+            `to_char(${dateFieldName},'DD-MM-YYYY') = '${parsedDay}-${parsedMonth}-${year}'`
+        ),
+      },
+    });
+
+    return appointments
+  }
+  public async findAllInMonthFromProvider({
+    month,
+    provider_id,
+    year,
+  }: IFindAllInMonthFromProviderDTO): Promise<Appointments[]> {
+    let parsedMonth = String(month).padStart(2, "0");
+    let appointments = await this.ormRepository.find({
+      where: {
+        provider_id,
+        date: Raw(
+          (dateFieldName) =>
+            `to_char(${dateFieldName},'MM-YYYY') = '${parsedMonth}-${year}'`
+        ),
+      },
+    });
+
+    return appointments
+  }
+
   public async findByDate(parsedDate: Date): Promise<Appointments | undefined> {
     let findAppointments = await this.ormRepository.findOne({
       where: {
@@ -28,14 +70,16 @@ class AppointmentsRepository implements IAppointmentsRepository {
 
   public async create({
     date,
+    user_id,
     provider_id,
   }: ICreateAppointmentsDTO): Promise<Appointments> {
     let appointment = this.ormRepository.create({
       provider_id,
+      user_id,
       date,
     });
-    await this.ormRepository.save(appointment)
-    return appointment
+    await this.ormRepository.save(appointment);
+    return appointment;
   }
 }