+++ Semana_4/node/src/modules/appointments/services/CreateAppointmentService.ts
@@ -1,26 +1,37 @@
 import Appointments from "../infra/Typeorm/entities/Appointments";
-import AppError from '@shared/errors/AppError'
+import AppError from "@shared/errors/AppError";
 import IAppointmentsRepository from "../repositories/IAppointmentsRepository";
-import {injectable,inject} from 'tsyringe'
-
+import { injectable, inject } from "tsyringe";
+import { isBefore, getHours } from "date-fns";
 
 interface IRequestDTO {
+  user_id: string;
   provider_id: string;
   parsedDate: Date;
 }
 
 @injectable()
 class CreateAppointmentService {
-
-
   constructor(
     @inject("AppointmentsRepository")
     private appointmentsRepository: IAppointmentsRepository // Cria automaticamente a variavel dentro da classe
-  ){
+  ) {}
 
-  }
+  public async execute({
+    parsedDate,
+    user_id,
+    provider_id,
+  }: IRequestDTO): Promise<Appointments> {
+    if (isBefore(parsedDate, Date.now()))
+      throw new AppError("This appointment could not be set to past date.");
 
-  public async execute({ parsedDate, provider_id }: IRequestDTO): Promise<Appointments> {
+    if (user_id === provider_id)
+      throw new AppError("You cannot book to yourself!");
+
+    if (getHours(parsedDate) < 8 || getHours(parsedDate) > 17)
+      throw new AppError(
+        "You can only create appointment between 8am and 5pm!"
+      );
 
     let findAppointmentInSameDate = await this.appointmentsRepository.findByDate(
       parsedDate
@@ -32,10 +43,10 @@ class CreateAppointmentService {
 
     let appointment = await this.appointmentsRepository.create({
       provider_id,
+      user_id,
       date: parsedDate,
     });
 
-    
     return appointment;
   } // Esse metodo deve ser unico no service
 }