+++ Semana_4/node/src/modules/users/repositories/IUsersRepository.ts
@@ -1,5 +1,6 @@
 import Users from "@users/infra/Typeorm/entities/Users";
 import ICreateUserDTO from "@users/Dtos/ICreateUserDTO";
+import IFindAllProvidersDTO from "@users/Dtos/IFindAllProvidersDTO";
 
 export default interface IUsersRepository {
   findByEmail(email: string): Promise<Users | undefined>; // encontrar por email
@@ -7,4 +8,5 @@ export default interface IUsersRepository {
   findByID(id: string): Promise<Users | undefined>; // encontrar por id
   save(user: Users): Promise<Users>; // save (atualziar)
   findAll(): Promise<Users[]>;
+  findAllProviders({except_user_id}: IFindAllProvidersDTO): Promise<Users[]>
 }