+++ Semana_4/node/src/shared/infra/http/routes/index.ts
@@ -1,5 +1,6 @@
 import { Router, json } from "express";
 import appointmentsRouter from "@modules/appointments/infra/http/routes/appointments.routes";
+import providersRouter from "@modules/appointments/infra/http/routes/providers.routes";
 import usersRouter from "@users/infra/http/routes/users.routes";
 import profileRouter from "@users/infra/http/routes/profile.routes";
 import sessionsRouter from "@users/infra/http/routes/sessions.routes";
@@ -15,6 +16,7 @@ routes.use("/appointments", appointmentsRouter); // Repassa qualquer rota http q
 routes.use("/users", usersRouter); // Repassa qualquer rota http que ira passar pelo caminho /users para o router de users
 routes.use("/sessions", sessionsRouter);
 routes.use("/password", passwordRouter);
-routes.use('/profile', profileRouter)
+routes.use("/profile", profileRouter);
+routes.use("/providers", providersRouter);
 
 export default routes;